refactor: 优化词云图代码
